Refactor types file name
