+++ src/components/search/SearchResult.vue
@@ -13,7 +13,7 @@
 
 <script lang="ts">
 import { Vue, Component, Prop } from "vue-property-decorator";
-import { Product } from "../../types";
+import { Product } from "../../types/model";
 import ProductItem from "./result/ProductItem.vue";
 @Component({
   components: { ProductItem },
